[105] Rename view component specs
We wanted to ensure that view spec and view previews mirror `app/component`
structure using sidecar pattern.  Each component will have a main folder
named after it. Within that folder we will ensure that there is a view spec
to test that functionality of the component and a preview file so that
developers can view components in the browser in specific states.