[41] Add Dttp::ContactService::Create
Add basic command class that creates a contact in DTTP.

This is a simple MVP class that just uses first_name, last_name and
persists the contactid that is returned as `Trainee#dttp_id`.
The intention is that we will expand this class and introduce a
presenter to translate between our domain `Trainee` and the DTTP
`contact` entity.

Co-authored-by: Sam Culley <samuel.culley@digital.education.gov.uk>