[41] Add dttp_id to Trainee
Add a UUID id to the Trainee model.

When the value has been set it is 'locked' as it should not be changed
after creation. It is still possible to change it obvs but this should
convey the intention that it isn't changed.

Each entity that we create in DTTP is assigned a unique ID.

Co-authored-by: Sam Culley <samuel.culley@digital.education.gov.uk>