[162] Use new TaskList component on the show page
The overview page should now match the overview page in the prototype.

I skipped these failing specs for the time being as its very likely that
more of the "Add trainee" journey is going to completely change. it
seems pointless to resolve these issues until we have more of the new pages
built.