[245] Add Autocomplete view component
The idea with this component is that it would be used within a form.
Instead of the component trying to build out the select list and include
all the styling for the default select list, it was decided that its best
to pass in a string of html. A div is wrapped around the passed in html
string and returned to the view.

Any autocomplete config options you would like to apply to the component
can be added by using the `classes` or `html_attributes` parameters.

`classes` takes a string which would be added to the parent div container.

`html_attributes` takes a hash which gets added to the parent div container

At the moment the only autocomplete options supported are:
- defaultValue
- showAllValues

It should be easy to add extra autocomplete options by adding the
option lookup to script.js and adding it to the autocomplete init
function call in script.js

No JS tests have been included in this spike because we need to setup
Jest and configure it to use webpacker because running the js unit tests.
This will be added in the future but in the meantime the component can
be manually tested using the preview components.