[504] Suppress the "Add another degree" button
In some instances we might not want the degree component to give the user
and option to add additional degrees. This commit adds a `show_add_another_degree_button`
param. By default this is set to true, so that it works without having to worry
about trying to update the sections confirmation page which is generic
page which other sections use but those section confirmation components will
not have this extra param.