[548] Remove "Error" from route output
initially, the idea was to playback some sort of Error
message back to the developer that there was something wrong
with the passed in data. After talking this through with Sam,
it would seem like this component should worry about whether or
not a route has been selected. The trainee record model should
control whether there is a record type or not. It should
never get to a stage where the record_type is missing from the trainee