[773] Remove rails-erd gem
After upgrading to Rails to 6.1 we got the following error from the rails-erd gem after running `rake db:migrate`:

rake aborted!
NoMethodError: undefined method `parent' for RegisterTraineeTeachers::Application:Class
Did you mean?  present?

The team has also agreed that it's not needed at the moment and should only be reintroduced if there's a good case for it and it supports the latest version of rails.