[813] Add Sluggable concern
Concern to add slug functionality to a model.

Add `has_secure_token` to generate a 24 character slug. This is the
rails default but we can reduce the length as the `has_secure_token`
does uniqueness checking.

Add `to_param` and `from_param` to allow use of the slug in URLs and
paths.