[872] Change not_disabled enum to no_disability
Enums create a negative scope by prepending `not_` to the enum value.
Creating an enum with `not_` as the prefix when there is also a positive
version  causes a warning that it will overwrite the scope.

```
2021-01-14 15:44:38.894378 W [1452:8180] Trainee -- Creating scope
:not_disabled. Overwriting existing method Trainee.not_disabled.
2021-01-14 15:44:38.894522 W [1452:8180] Trainee -- Enum element
'not_disabled' in Trainee uses the prefix 'not_'. This has caused a
conflict with auto generated negative scopes. Avoid using enum elements
starting with 'not' where the positive form is also an element.
```