[888] Add active links to navigation
Add active styling (blue bar) to navigation links.

Links are active if they are explicitly marked as current OR the link
matches the current path.

Refactor to remove coupling with ActionView::Helpers::UrlHelper. Pass in current_path
when instanciating the component.

Remove authenticated check to reduce coupling. This check is now made in
the view when instanciating the component.

Co-authored-by: Avin Hurry <avin.hurry@digital.education.gov.uk>