[961] Move status update queueing
* Move the enqueueing of the status jobs out of `Dttp::RegisterForTrn`

The `RegisterForTrn` service shouldn't be coupled to being in an async
environment. Moving these calls into the register job removes this
coupling.

It is unfortunate that we probably 'always' want to to set the state
when registering for TRN. We could maybe rename this service 'create
contact' or something like that and have a separate service that does
the three operations requried for registration. The problem with that is
that if one of the status operations fails we might end up retrying the
creation and spamming DTTP with lots of duplicates. This approach seems
like a reasonable compromise.