[961] Add programme year and length
Hardcoded to 1 for now.

Programme year is the year of programme length.

e.g

Year 1 of 2 is the first year of a 2 year course.

Year 2 of 1 could mean a trainee that is still on a 1 year course after
the date of census. Maybe they hadn't been awarded QTS as yet.