[1207] Revert "Make captialisation consistent"
This reverts commit b01c047d6a6b9b53ae31a3f95ffc9736590161ad.

We can't make any changes to the keys now that there are degree records already in prod with matching `grade` values. If we change them, then `CodeSets::Grades::MAPPING.dig(grade, :entity_id)` will not be able find to the matching value. This coupling is not ideal, so we may migrate Degree#grade to enum type later on.