[1243] Add partial validation to autocompletes
Adds a validation to all autocompletes (excluding country) so that the
User will get a specific message if they type something into the field
but fail to select a value.

This works by submitting a text and select input for the autocomplete
field and comparing the two in a form object validation.

To test this, it was easiest to do a feature test because there are 
parts in the views, controller and form objects that need to all slot
together to ensure the validation is correctly shown to the user. This
required setting up chrome web driver on CI, and enabling js for these
tests using `js: true`

The country autocomplete was left out because it uses an external js
library which would be hard to override, and it doesn’t suffer from the
same usability problem because it uses the “autoselect” option for the
Accessible-autocomplete library.