[1306] PoC for giving specific error messages for partial autocompletes
The goal here is handling the case where a user types something into an
autocomplete field but doesn’t select anything from the resulting
dropdown. Existing behaviour will either ignore the user’s input and
keep the field filled with the previous value, or just blank it out,
which can be confusing.

We want to retain what the user entered into the text field when
re-rendering the page with errors, and provide a specific error message
for that field being partially submitted.

This approach validates that the text input is in a list of the options
that were presented to the user. This should be more resilient to any
type of autocomplete we want, but the implementation seems very messy
as it requires reassigning all the errors after the form object has
been validated, see code comment for more details.