[1311] Add publish course details page
Adds a feature flag that will direct the user to a new page where they
can select from a list of courses (or choose to enter manually).

List of courses shown will be based off the route/provider, but this
Is not yet ready, so just pick 10 random courses for now

If the user selects a course from the list then send them to a confirm
page (different from existing course details confirm), where the
course attributes will be copied to the trainee.