[1311] Allow task list to have callable paths
This tidies up the trainee helper a bit, and allows us to remove the
stashed_code method.

The reason why we needed that is because it isn’t possible to generate a
valid confirm_path before the user has started. Using a lambda instead
allows us to just use the correct route helper and not worry about it 
being generated prematurely