[1414] Add RetrieveTrnJob.perform_with_default_delay
* Refactor RetreiveTrnJob to encapsulate the default delay and polling
  limit configuration and update the tests to check that the
configuration is used rather than the literal values are used.

The config delay values are configurable so we shouldn't test explicitly
for the currently configured value or otherwise all of the tests break
if we change the configuration.

They were stored on constants in the class that were then passed in
again as arguments when the job was queued. This meant the calling code
needed to know about them.