[1630] Add searchable column to schools
This column stores a tsvector that we can index for using in searches,
giving a big performance boost.

We need to update this tsvector column any time the school changes so
that the search index can be kept up to date. This must be done in an
active record callback (alternatively using a database trigger would
require using some raw sql in a migration that couldn’t be represented
in schema.rb, and we don’t want to use an sql formatted schema).

Also add an index on the close_date as we can search for only open 
schools, this also speeds up the search query.