[1243] Fix autocomplete validation for additional nationalities
The main issue with this was that the selects were using ids as the
values of the select, meaning the autocomplete validator was always 
detecting a mismatch no matter what the user typed:

- Convert the nationality selects (and nationality checkboxes) to use
names as the values, convert to/from these on the backend to ids.

- Exclude raw params from saving to the trainee (was also causing an
Error)

- Modify js for nationality select to allow for errors on nationality
fields. This was a bug that would previously prevent remove links and
other functionality from working if there was an error on any of the
nationality fields because the form builder adds “-error” to labels and
ids of fields that have errors