Add custom trigger to handle postcodes
One of the problems with the previous dictionary is that postcodes
would be indexed as two separate words. If you were to type in a
postcode without spaces it wouldn’t return the result you wanted. To
Fix this we need to add the postcode without the space into our tsvector.

The search performance seems unaffected (this difference between 1-2ms/
0.09ms execution time seems to fluctuate, possibly depends if the query
gets cached or not. But now we can see the full postcode has also been
Indexed

irb(main):017:0> School.first.searchable
=> "'100000':1 '5de':6 'aldgate':3 'ec3a':5 'ec3a5de':7 'london':8 'school':4"