[1652] Add apply trainee data page
This page is similar to the check details page in that there will be
Subsections of data that come from an apply application. The difference
being that in a future iteration, there will be error validation that
will point the user to fields they need to correct, rather than just
being told a section is incomplete and having to navigate through
every step of the form.

Given the similarities with the check details page (and trn submission
Form), I tried to share some behaviour between them, but decided to go
for a bit of copying until we add more functionality, namely:
- Error checking on the apply trainee data page
- Apply draft version of the check details page