Fix flakey employing school search spec
This fix has already been applied to the lead school search spec.

The click method switches between using `.click` and `.trigger(:click)`
depending on if the test is using js or not. `.trigger(:click)` needs
to be used when we are using js as capybara can start running before
the web driver has initialised its event hooks (eg. Triggering the click
event handlers when something is clicked), so we need to trigger it
manually.