[1652] Handle unanswered diversity questions
Since users can potentially not have any diversity questions answered
now when we are showing the info in the apply-trainee-data page, there
Was a bug where having a diversity_disclosure: nil would render
```
{:diversity_disclosed=>”Information disclosed.”
 :diversity_not_disclosed=>”Not shared.”}
```

These are actually the two translation keys for the two options in the
Yaml file.

To fix this, add a new key to show that the information has not been
provided and say “Not provided” similar to all the other sections on 
this page. Eventually there will also be an error message to the user
to prompt them to fill this information out.