[1652] Collapse education section when not started
When this section has no information, we still want to have it
accessible on the page. A user could get into this state by deleting
a degree when the section is expanded.

Change terminology of complete vs incomplete sections to be 
expanded/collapsed as we will now be showing expanded incomplete
sections on the apply trainee data page.