[1744] Refactor RetrieveTrnJob
Make the timeout date an argument so as we can requeue the job after it
has timed out.

We could do this previously without changing the submitted at date on
the trainee as the job used that to work out if it had timed out.

Now we can set it with the default timeout using the helper method or
queue it with a custom date.