[1899] Handle DTTP users with nil email
There are a number of users in DTTP with no email address. This is
causing the provider#show page to error as we don't handle it. These
users should not be imported into register without an email so don't
display them.

The operations team will review the other users and sync should update
them. There are currently 11.