[1849] Re-add training route filtering for trainees
Also break the filtering options out of the component. Since we want to
only show routes that there are trainees for, this would require a 
db query. It seemed wrong to inject this set of options all the way down
from the controller through the paginated filter -> filter (and also
to not do this with other options, like status and subject).