[1988] Refactor confirm details routes
The helper for calculating the keys for progress relied on the existence
of a path that had the progress key at the end. For the degrees section
this was weird as the trainee_degrees_path wasn’t actually used for
anything other than this (no index action in the degrees controller).

For funding we don’t want to have to create a path that isn’t used 
anywhere just so we can look up the progress key, so change this helper
to find