[2225] Add initial hoppit csv import task
Takes a csv of trainee data and creates trainees out of it. Runs as many
validations as we can to prevent trouble when communicating with DTTP.

If any rows are invalid, fail the whole task and don’t create any 
trainees so we don’t have to worry about idempotency.