[2116] Reset update SHAs.
* Reset the `Trainee#dttp_update_sha` for all trainees that have not
  been updated in the last 24 hours.

We've changed the algorithm for creating the Trainee#dttp_update_sha
which will make all trainees appear like they need an update.

Reset any that have not been recently updated. This will likely cause
some trainees to have an uneccessary update sent to DTTP but will
reduce it to a reasonable level.