[2314] Fix contact update error
We already have a feature flag to disable this job in the review 
environment because the generated trainees don’t have complete data.

There is an error occurring in the initialiser for this service despite
It being disabled, so add another check in the initializer to short
circuit even earlier.