[2376] Don’t send degree date obtained as part of degree payload to dttp
This field doesn’t exist, and is causing a 400 error when we try to 
send it to dttp. The degree qualifications model only has a degree
end date as a possible value rather than the graduation date, but it 
seems to be null all the time. The degree date obtained is still set
on the placement assignment

We also need to change the format of the degree date that we send 
On the placement assignment, this was previously parsing as the 1st of
October. Odata DateTimeOffset (the type of the field) requires the 
value to be in ISO 8601 format