[xxx] Default show_funding to true
The show_funding feature is on now in production. Set it to default on
in the base settings.yml to ensure it is enabled in all environments.

We can remove this feature flag now but are about to start work on
moving them to the database so we can do it as part of that work.