[2870] Use the interstitial confirmation page when invalid degrees exist
Now, since we have data coming in from various sources, for example, the
apply import, or via csv import, it is possible that a trainee has degree(s)
which might be invalid. The current flow from the review-draft page is
to take the user to a form to add a new degree (as opposed to enabling
them to edit an existing degree.

This change will ensure that users are redirected to the degrees confirm
page, when some degree data is avaiblable against that trainee, so that
the user can decide to amend the existing degree information, as opposed
to being forced to add a new degree.