[xxx] Don't retry update jobs
* Set retry count to 0 on Dttp::UpdateTraineeToDttpJob

Update jobs are requeued every evening if the trainee still has unsent
changes. They are also very flakey at the moment as we're making records
invalid in the console when we change route.

This change will make the jobs go straight to the dead set rather than
build up retries and spam sentry. This should make them more manageable
for support until we fix this feature.