Create terraform tfvars.json files
Instead of using tfvars which is proprietary terraform format, we use
json which can easily be consumed by standard tools like jq. jq is even
available by default on the Github actions runners.

The keyvault configuration was stored separately in .sh files so they
could be consumed by terraform and the workflows. It's now moved into
json as well and consumed by terraform and jq in the workflows.