[3107] Reimport apply applications changed since 09 Dec 2020 14:00
Some of the apply applications imported into register are missing the
course uuid that we added in recently. To ensure that we re-import these
applications, we want to trigger the import job by passing in the
changed_since parameter to the apply request.

The earliest changed_since/updated_at value found in our database is
Wed, 09 Dec 2020 14:23:24.000000000 UTC +00:00.
Hence this data migration will ensure that we fetch all of applications
added/changed since.