Handle multiple placement assignments
Trainees could get a new placement assignment in various scenarios,
such as: a) if they are still active at clockover or b) if they change course mid cycle.

For this import we are fetching the everything from the last PA ordered by date,
except the the trainee's start date which is fetched from the first one.