[xxx] Remove ERD gem
This gem was being used to generate a database diagram with every
migration. This often causes issues for the developers during rebasing
etc where the diagram gets out of sync or updated when it shouldn't.

The overhead isn't worth the dev time when the schema is in constant
flux. We also have the schema in a db/schema.rb. If we really need a
diagram it's a 2 minute dev task to generate one from main.