[3452] Add UserWithOrganisationContext
* Add UserWithOrganisationContext decorator class

This class takes the current user and the session and encapsulates
behaviour to expect the current organisation throughout the request
cycle. An instance will be returned from
`ApplicationController#current_user` and enable the app to determine
which organisational 'hat' the user is currently wearing.

The `#organisation` implentation here is a placeholder to allow us to
incrementally move towards the full implementation. In the future it
will return the provider or lead school that the user has chosen to work
as. Currently it returns `providers.first`.