[3508] Update TraineePolicy
* Add support for lead school user context
* Permissions are now either read? or write?

Lead school users only have read access, provider users (and system
admins) have read and write.

The spec is a bit tricky for this as we've mixed the concerns a bit by
adding the trainee state transition logic into the policy. Add an
exclusion to rubocop-todo as it didn't like the specs. Not sure if it
understands the `permission` blocks as describe/context.