[3574] Save ethnic_background as Not provided for white ethnicities
If the DTTP ethnicity is White or Scottish, then we save ethnic_group
as White.
But we also need to save ethnic_background to Not provided, otherwise
the section is incomplete.
I've also fallen back on Not provided for both ethnic_group and
ethnic_background if the DTTP ethnicities is completely missing.