[3508] Remove `User#primary_provider`
* remove the primary_provider method and favour
  current_user.organisation
* Update the system_admin routes to remove the shallow: true from the
  nested provider user paths. We need the provider id in these controllers.

Users don't have a 'primary' provider. This was temporary as we switched
from a belongs_to to a has_many stylee relationship.

In most cases the provider is retreived from the
current_user.organisation object. Where the user only has one provider
this will always return the provider. Where the user has multiple then
the current context is retreived from the session. Most of the existing
specs assume only one provider so `current_user.organisation` can be
substitued.

Some specs had to be reworked to accomodate this change.

The new format provider-user routes mean that delete user is currently
only exposed on the provider. That can be updated in the future.