Map STANDARDS_MET to recommended_for_award or trn received
For hesa records, we need to map STANDARDS_MET to trn_received, as they
do not use this status to signify that a tranee has been recommended for
award.

For portal records, however, we need to rely on the value of
dfe_recommendtraineetonctl, which is set to true when a record is
actually recommended for award. This change will use the value of this
boolean to determine if the register status should map to
recommended_for_award