[3603] Split TraineePolicy create? and new? methods
To be able to render the new trainee buttons,
we do not need to match a provider with a trainee, but on the create?
policy, it is safer to ensure that the user creating the trainee is
associated to the same provider. To ensure this, we separate out the
new? and create? policy checks.