[3743] Check permission for CSV
Add an authorization check to the CSV download of trainees.

Some users (lead school users currently) aren't authorized to download
the trainee CSV. Check the `Trainee#export?` policy before allowing.

Now we are doing an authorization check on a request for a .csv, Rails
looks for the forbidden.csv template if we don't specify an extension on
the redirect in ApplicationController where we catch the Pundit error.
Adding a CSV formatted response isn't useful for the users and putting
HTML in a .csv file doesn't seem very good either so specify the
extension on the redirect to force rails to render the HTML. This also
required a bit of a change in the locale file.