Avoid OpenStruct
The reasons for not using this seem to be down to poor performance. It
is really convenient and we've used it in places like specs and
initializers where we probably don't care. So...

Exclude spec, db and config directories
Replace OpenStruct with classes where it is used in the app