[3778] Fix study mode mapping
`CR_MODE` was previously mapped to HEI Census Status in DTTP. It indicates the mode the student was reported on for the DfE census in their first year only.

Historically, there was obviously a need to know this but it's not relevant any more and the standard mode we capture in Register now suffices.