Stop deployment status marking deployment inactive
The deployment status update task by default (in version 1.1.0) actively
marks previous deployments as inactive, this results in a large number
of API calls which in turn causes a rate limiting failure.  This change
disables that function, this will leave deployments showing as Active in
the deployments page.  An issue has been raised to fix this in the
action.