Fix check on smoke test outcome
The wait for step wasn't locating the dispatched smoke test workflow so
was timing out.  This wasn't failing the workflow because the condition
only checked for failures and not timeouts.  Changed the ref used to
track the dispatched workflow and altered the condition to fail the
workflow if smoke tests not successful.