[3839] Fix undefined method 'recruitment_cycle_year' for nil:NilClass
Fixes https://sentry.io/organizations/dfe-teacher-services/issues/2967214933

This was getting triggered when a user tries to create a trainee on a course
where previously the full time/part time dates have been set globally.

The exception was being triggered due to the course_uuid not being sent
through to the form.