[3899] Initialize dqt_update_sha
Set dqt_update_sha in a data migration.

We determine if a record needs to be sent as an update by setting a SHA
of the record and then checking if it has changed. We need to initialize
this before we turn on the DQT integration to avoid sending unnecessary
updates.