[3905] Add Trainees#cohort
An enum that indicates whether we consider this record current, past or
future. This will be set on update and by a nightly process to
transition trainees to past when the academic year changes.

That logic will be in a separate service but persisting the state on the
trainee will simplify filtering.