Explicitly set RAILS_ENV to avoid GitHub test build from failing
Sometimes the test build fails with the following error:

Web Console is activated in the test environment. This is
usually a mistake. To ensure it's only activated in development
mode, move it to the development group of your Gemfile:

    gem 'web-console', group: :development

If you still want to run it in the test environment (and know
what you are doing), put this in your Rails application
configuration: