[3976] Add Trainees::Update
A service that:

* updates the trainee
* queues a Dqt::UpdateTaineeJob
* queues a Trainees::SetCohortJob

We have used after_commit callbacks for this but the cohort job causes a
sidekiq loop as it requeues itself and this approach is more explicit
and means we can update the trainee without queueing all of the jobs
too.

In some places in the app we set attributes on the trainee prior to
update rather than passing params to update so support that too.