[4055] do gov.uk redirect in routes
This was previously being done as a part of Rack middleware. While that
might be slightly more efficient, it's likely to be more surprising to
any devs workind on the project. And also testing is likely to be more
tricky.

This commit moves the redirect into the routes file alongside the
traineeteacherportal redirect so it's a bit clearer. We also take the
opporunity to introduce a feature flag to be able to control this better.