[4129] Build a local HESA data cache
We've identified issues and inconsistencies with HESA data, some of which involve data imported to DTTP prior to our April update. At some point we'll most likely need to use data submitted during census and the January update to fix issues.

Running a full import over the XML is very slow and the XML data is difficult to use for analysis which we need to both fix the current issues and more easily identify others.

We have used a pattern with DTTP where we store the source data in the register database which makes fixes fast and analysis much simpler.