Change auth rules: allow admins to see drafts
Previously we were hiding the menu item and denying access to _Drafts_
if a user has a lead school even if they were system admins.

This change permits system admins to always see Drafts. It also changes
the menu rendering check to refer directly to the policy rather than via
a helper so that the rules and front-end are easier to keep in sync.