Fix flakey spec
I got an off by one second error here because the code under test runs
before we use Timecop to freeze time so if you are very unlucky the
clock will tick a second between the code running and the freeze.

Solution here is to just move the freeze to an `around` block that
should run before the `before` block.