[4160] Allow Trainees::CreateFromDttp to use a custom placement assignment
This will help us create a duplicate trainee but for a different placement assignment. This situtation arises when a provider has a trainee that withdrew from a course and has subsequently started a different course. The provider wants a record of the trainee in both states and the only way at the time of writing to represent this in register is duplicate records.