[4250] Refactor `trainees_awarded` tile
* Remove the `Trainee.awarded_in_cycle` 'scope'
* Refactor the `HomeView` tile to directly use the academic cycle scope

The existing scope was buggy as not being scoped to the trainee instance
it was on, it didn't chain properly with others. It was also a weird
crossover of dependency between `Trainee` and `AcademicCycle`

Merge the `AcademicCycle.trainees_ending` scope in the view to avoid
this and fix the bug.