[4352] Add Degrees::FixToMatchReferenceData service
* Add service

This service takes a degree and attempts to find a match for each of:

type
institution
grade
subject

from the DfE::ReferenceData. If it finds a match, using case insensitive
matching or the synonyms in the data, then it updates the degree with
the new matching value.

We haven't updated the UUIDs here yet as this fixes a production
issue that is affecting users. There is work scheduled to backfill the
UUIDs.