[4369] Fix the logic to determine education phase and primary cours attributes
It's more accurate to use the course age range to determine the course education phase. We've been having cases where a trainee is doing primary teaching but the course education phase is marked secondary.

We also encounter situations where a primary course has "primary teaching" missing in the subjects or "primary teaching" in the second subject when it should be the first subject store in Register.