[xxx] Disable welcome email
* Set send_email feature flag to false in all environments

We are bulk importing lead school users and have an email to send them
to tell them that they have been added to the system.

We don't want to email them again with the welcome email which is
currently sent after they first login.

Disable it for now and we'll decide if we get rid of it altogether or
send it at a different point.