[4475] Remove min_age and max_age validation from Course model
Publish does not currently validate age ranges within the context of course level (primary or secondary). This means we can receive courses that are for primary level but the max age is greater than 11 which Register considers secondary level. Therefore, if a published course is primary level and the max age is greater than 11, we don't store the age range. Instead, when the providers selects the course, the age range will have to entered manually. This ensures the age is valid against Register's age range validation rules.