[4502] Add collection_reference column to hesa_students table
We're not sure how long the hesa_students table will be around for. It has proved to be very useful in backfilling missing data. With the C21053 collection coming to a close, we want to backup the whole collection from the start of the academic year just in case we need it for future reference. Adding a collection_reference column will help us know the source of each record, especially if we intend on store future collections.