[4897] Increase production app instances to 4
Production has been going down due to excessive load on expensive pages such as /trainees. After load testing the site, it was determined that 2 app instances is not enough especially during busy periods such as Oct-Nov. 4 app instances proved to be much more performant and was able to handle heavy load. Autoscaling is probably the long term approach, but for now, adding more resources gives the users a better experience and reduce downtime.