[xxx] Fix HESA importer with regards to disabilities
This is the first year that HESA has supported multiple disabilities. They might not have thought to add validation that some of the options are mutually exclusive. For example, one trainee had "Mental health condition" and "No known disability" which shouldn't be possible. "No known disability" is not in the database of disabilities therefore Disability.find_by(name: disability) return nil and that led to a ActiveRecord error AssociationTypeMismatch during the save operation. Using `compact` will ensure nil values are removed, but hopefully HESA will implement validation in the future.