[4912] Update DQT after HESA trainee import
We had a background job Dqt::SyncStatesBatchJob that sent updates to DQT every day, but it was taken out of operation because it was transitioning some trainees to awarded when they were still in training. If a trainee previously had QTS and is now training for EYTS, the DQT v1 API returns ‘awarded’ - it doesn’t distinguish between the two qualifications.

DQT are working on v2 which we think returns better data, but we don’t want to turn Dqt::SyncStatesBatchJob back on until we’ve done the work to use the new v2 API. For now, we'll send the DQT update after the trainee is imported from HESA.