Add placement details to manual entry guidance (#2933)
* Add placement details to manual entry guidance

We want providers to know that they will be asked about placements later in the academic year, so they should keep a record even though they can't add the data now. I've added a section to the page of guidance about data needed to manually register a trainee.

I've used inset text to match the diversity data section. This is information about the data requirement, rather than a list of data items as in most sections.

![CleanShot 2022-11-15 at 17 58 44](https://user-images.githubusercontent.com/29047487/202485849-405520b1-070c-4341-93cc-b5c7a8837d63.png)

* Added <p> tags needed within <div> for inset text

I thought that I could omit the <p> tags within the <div> for inset text, since this is markdown. But that combined the two paragraphs into a single paragraph, so I've added the <p> tags.

* Update app/views/guidance/check_data.md

Co-authored-by: Ed Horsford <mail@edwardhorsford.com>