[5103] Set/recalculate trainee's academic cycles within Trainee instead of background job
We started to see some trainees with incorrect end years. After an investigation, it was postulated that because the academic years were updated using a background job in various places, it's possible that either the job failed or it was never executed during a manual update.

To ensure the academic years is always correct, we have decided to move this calculation into the Trainee's model and executed after every save.