[xxx] Fix Dqt::WithdrawTraineeJob
Using Trainee#submitted_for_trn_at to calculate the timeout was not a good idea since it sometimes can be nil or it's too far in the past to be useful. This job can we kicked off anytime in the future, so we just need to terminate it after a number of days if unsuccessful.