[xxx] Cancel withdraw job if trainee is already withdrawn in DQT
We're getting a lot of dead withdraw jobs because the trainee is already withdrawn in DQT. Not sure how the trainee got to be withdrawn before Register knew about it, but for now, this workaround will reduce the workload on support.

The trade-off is the withdraw job may end up doing 2 API calls to DQT which may impact the 3000/per minute rate limit during busy periods.