Updated 'request account' page
The content of the 'request account' page said that people should contact us to request their own account. But if their organisation already uses Register then they should ask a colleague with an account to contact us. This is now the most common situation.

I've taken the opportunity to also make minor content changes throughout.

In addition, this page was referring to content in /config/locales/en.yml. That made it difficult to update and was unnecessary since we do not have other language versions. I've replaced these references with html. In future we may change the page so that it uses markdown.